Fix paring in x & y in G2
Looks stupid at first, but, for some other reason, that I do not know
we needed to encode the points in a reverse order, the coordinates of
the point itself not in the (c0, c1) but in the (c1, c0), silly I know.
not what I expected and not what e studied at school lol.